refactor: group codegen specs
