+++ package.json
@@ -97,7 +97,7 @@
   "codegenConfig": {
     "name": "RNCWebViewSpec",
     "type": "all",
-    "jsSrcsDir": "./src",
+    "jsSrcsDir": "./src/specs",
     "android": {
       "javaPackageName": "com.reactnativecommunity.webview"
     },
