+++ src/WebView.ios.tsx
@@ -2,8 +2,8 @@ import React, { forwardRef, useCallback, useImperativeHandle, useRef } from 'rea
 import { Image, View, ImageSourcePropType, HostComponent } from 'react-native';
 import invariant from 'invariant';
 
-import RNCWebView, { Commands, NativeProps } from './RNCWebViewAppleNativeComponent';
-import RNCWebViewModule from './NativeRNCWebViewModule';
+import RNCWebView, { Commands, NativeProps } from './specs/RNCWebViewAppleNativeComponent';
+import RNCWebViewModule from './specs/NativeRNCWebViewModule';
 
 import {
   defaultOriginWhitelist,
